+++ examples/shared/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @vetojs-examples/shared
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [4be5eeb]
+- Updated dependencies [b2e7ab2]
+- Updated dependencies [4be5eeb]
+  - @vetojs/core@0.10.0
+
 ## 0.0.9
 
 ### Patch Changes